Make compatible with Excel file & iterate through rows

+++ product/product_actions.py
@@ -5,9 +5,11 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver import Firefox
 import requests
 import time
+import math
 import os
 
 def enter_description(driver: Firefox, wait: WebDriverWait, description: str):
+    wait.until(EC.element_to_be_clickable((By.XPATH, "/html/body/div[1]/div[4]/div[2]/div[1]/div[1]/div[7]/div[3]/a")))
     driver.find_element(By.XPATH, '/html/body/div[1]/div[4]/div[2]/div[1]/div[1]/div[7]/div[3]/a').click()
     wait.until(EC.element_to_be_clickable((By.XPATH, "/html/body/div[1]/div[4]/div[2]/div[2]/div/div/div[2]/div/div[2]/div[1]/div[2]/div[1]/iframe")))
     descriptionIframe = driver.find_element(By.XPATH, '/html/body/div[1]/div[4]/div[2]/div[2]/div/div/div[2]/div/div[2]/div[1]/div[2]/div[1]/iframe')
@@ -35,15 +37,18 @@ def set_google_category(driver: Firefox, wait: WebDriverWait, category: str):
 def open_find_and_filter_options(driver: Firefox):
     driver.find_element(By.XPATH, '//*[@id="app"]/div[4]/div[2]/div[1]/div[1]/div[13]/div/div[10]').click()
 
-def set_primary_colour(driver: Firefox, image_url: str):
-    file_path = "product_images\\" + image_url.rsplit("/", 1)[-1]
-    colour = find_colour_for_image(file_path)
-    print(colour)
+def set_primary_colour(driver: Firefox, primary_colour, image_url: str):
+
+    if (math.isnan(primary_colour)):
+        file_path = "product_images\\" + image_url.rsplit("/", 1)[-1]
+        colour = find_colour_for_image(file_path)
+    else:
+        colour = primary_colour
     driver.find_element(By.XPATH, '/html/body/div/div[4]/div[2]/div[1]/div[1]/div[13]/div/div[10]/div/div/div/div/div/div[2]/ul[2]/li[1]/div[7]/div[3]/div[2]').click()
     driver.find_element(By.XPATH, f'/html/body/div/div[4]/div[2]/div[1]/div[1]/div[13]/div/div[10]/div/div/div/div/div/div[2]/ul[2]/li[1]/div[7]/div[3]/div[2]/div/div/div[@data-value="{colour}"]').click()
     driver.find_element(By.XPATH, '/html/body/div/div[4]/div[2]/div[1]/div[1]/div[13]/div/div[10]/div/div/div/div/div/div[3]/button').click()
 
-    if (os.path.isfile(file_path)):
+    if (math.isnan(primary_colour) and os.path.isfile(file_path)):
         os.remove(file_path)
 
 def block_sim_stock(driver: Firefox):
